requirements.txt
# src/requirements.txt

google-generativeai
requests
python-dotenv
mcp
fastapi
uvicorn
pytest
google-cloud-aiplatform